backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
geoalchemy2==0.17.0
pydantic==2.10.3
pydantic-settings==2.7.0
redis[hiredis]==5.2.1
httpx==0.28.1
standardwebhooks==1.0.1
python-multipart==0.0.20
aiobotocore[boto3]==2.15.2
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
shapely==2.0.6
fiona==1.10.1
python-jose[cryptography]==3.3.0
